refactor: add fs dir
