+++ src/services/fs/createEmptyFile.js
@@ -1,8 +1,8 @@
 import fs from "node:fs";
 
-import { normalizePathString } from "../utils/normalizePathString.js";
+import { normalizePathString } from "../../utils/normalizePathString.js";
 
-export const operationAddFile = async (comandKey, data) => {
+export const createEmptyFile = async (comandKey, data) => {
   const normalizePath = normalizePathString(comandKey, data);
 
   fs.open(normalizePath, "ax", (err, _) => {
