+++ src/services/index.js
@@ -2,8 +2,4 @@ export * from "./closeFileManager.js";
 export * from "./operationLs.js";
 export * from "./operationUp.js";
 export * from "./operationCd.js";
-export * from "./operationCat.js";
-export * from "./operationAddFile.js";
-export * from "./createDirectory.js";
-
-
+export * from "./fs/index.js";
